Done migration with EF

+++ API/Domain/Entities/Blog.cs
@@ -4,7 +4,7 @@
         public Guid Id { get; set; }
         public required string Title { get; set; }
         public required string Content { get; set; }
-        public required int PriceOfSubscription { get; set; }
+        public required double PriceOfSubscription { get; set; }
         public required List<Comment> Comments { get; set; }
     }
 }
